Required changes to make Kdenlive work with some locales that have a comma (,) as numeric separator. Also now documents with a different locale will open as read only. Sadly this breaks the string freeze by introducing a new string and probably messing up some effects translation, since I changed the "paramlist" type to use ';' as separator instead of ','.
svn path=/trunk/kdenlive/; revision=5793

+++ src/dvdwizard.cpp
@@ -305,9 +305,9 @@ void DvdWizard::generateDvd()
             // We need to make sure that the y coordinate is a multiple of 2, otherwise button may not be displayed
             buttonsTarget.append(it.key());
             but.setAttribute("x0", QString::number(r.x()));
-            but.setAttribute("y0", QString::number(2 * (r.y() / 2)));
+            but.setAttribute("y0", QString::number((int) 2 * (r.y() / 2)));
             but.setAttribute("x1", QString::number(r.right()));
-            but.setAttribute("y1", QString::number(2 * (r.bottom() / 2)));
+            but.setAttribute("y1", QString::number((int) 2 * (r.bottom() / 2)));
             spu.appendChild(but);
             i++;
         }
